requirements.txt
Flask==2.1.3
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.10.0
azure-keyvault-secrets==4.5.1
azure-storage-blob==12.13.1
requests==2.28.1
requests-oauthlib==1.3.1
tensorflow-cpu >= 2.9.1
numpy >= 1.23.1
opencv-python >=4.6.0.66
